+++ app.js
@@ -3,7 +3,7 @@ const logger = require('morgan')
 const cors = require('cors')
 require('dotenv').config()
 
-const authRouter = require('./routes/api/auth')
+const usersRouter = require('./routes/api/users')
 const contactsRouter = require('./routes/api/contacts')
 
 const app = express()
@@ -14,7 +14,7 @@ app.use(logger(formatsLogger))
 app.use(cors())
 app.use(express.json())
 
-app.use('/api/auth', authRouter)
+app.use('/api/users', usersRouter)
 app.use('/api/contacts', contactsRouter)
 
 app.use((req, res) => {
