+++ controllers/user/register.js
@@ -5,16 +5,18 @@ const { User } = require('../../models/user')
 const { RequestError } = require('../../helpers')
 
 const register = async (req, res) => {
-  const { name, email, password } = req.body
+  const { email, password } = req.body
   const user = await User.findOne({ email })
   if (user) {
     throw RequestError(409, 'Email in use')
   }
   const hashPassword = await bcrypt.hash(password, 10)
-  const result = await User.create({ name, email, password: hashPassword })
+  const result = await User.create({ email, password: hashPassword })
   res.status(201).json({
-    name: result.name,
-    email: result.email,
+    user: {
+      email: result.email,
+      subscription: 'starter',
+    },
   })
 }
