+++ models/user.js
@@ -7,24 +7,24 @@ const emailRegexp = /^[a-z0-9]+@[a-z]+\.[a-z]{2,3}$/
 
 const userSchema = new Schema(
   {
-    name: {
-      type: String,
-      required: true,
-    },
     email: {
       type: String,
-      match: emailRegexp,
+      required: [true, 'Email is required'],
       unique: true,
-      required: true,
     },
     password: {
       type: String,
       minlength: 6,
-      required: true,
+      required: [true, 'Password is required'],
+    },
+    subscription: {
+      type: String,
+      enum: ['starter', 'pro', 'business'],
+      default: 'starter',
     },
     token: {
       type: String,
-      default: '',
+      default: null,
     },
   },
   { versionKey: false, timestamps: true },
@@ -33,7 +33,6 @@ const userSchema = new Schema(
 userSchema.post('save', handleSaveErrors)
 
 const registerSchema = Joi.object({
-  name: Joi.string().required(),
   email: Joi.string().pattern(emailRegexp).required(),
   password: Joi.string().min(6).required(),
 })
